+++ crates/macros-next/src/lib.rs
@@ -2,7 +2,7 @@
 
 use proc_macro::TokenStream;
 use quote::quote;
-use schematic_core::container::Container;
+use schematic_core::container::{Container, ContainerMacro};
 use syn::{DeriveInput, parse_macro_input};
 
 // #[derive(Config)]
@@ -15,13 +15,24 @@ pub fn config(item: TokenStream) -> TokenStream {
     quote! { #output }.into()
 }
 
+// #[derive(ConfigEnum)]
+#[cfg(feature = "config")]
+#[proc_macro_derive(ConfigEnum, attributes(config, variant))]
+pub fn config_enum(item: TokenStream) -> TokenStream {
+    let input: DeriveInput = parse_macro_input!(item);
+    let mut output = Container::from(input);
+    output.macro_type = ContainerMacro::ConfigUnitEnum;
+
+    quote! { #output }.into()
+}
+
 // #[derive(Schematic)]
 #[cfg(feature = "schema")]
 #[proc_macro_derive(Schematic, attributes(schematic, schema))]
 pub fn schematic(item: TokenStream) -> TokenStream {
     let input: DeriveInput = parse_macro_input!(item);
     let mut output = Container::from(input);
-    output.schematic_only = true;
+    output.macro_type = ContainerMacro::Schematic;
 
     quote! { #output }.into()
 }